Refactor loading gui into a single module
