+++ alignment-software/alignment_software/gui/window_main.py
@@ -7,7 +7,7 @@ from alignment_software.engine.particle_tracking import (
 from .manual_track.step_manual_track import ManualTrackStep
 from .common import AsyncHandler
 from .contrast.step_contrast import ContrastStep
-from .loading.step_loading import LoadingStep
+from .loading import LoadingStep
 from .transform.step_transform import TransformStep
 from .coarse_align.step_coarse_align import CoarseAlignStep
 from .auto_track import AutoTrackStep
